% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecasting.prob.R
\name{Forecasting.prob}
\alias{Forecasting.prob}
\title{Creates Marginal Distributions in the Forecasting of Time series}
\usage{
Forecasting.prob(k, q)
}
\arguments{
\item{k}{number of categories}

\item{q}{The number of the domain in which the value of the independent variable is located}
}
\description{
Creates Marginal Distributions in the Forecasting of Time series
}
